backend/requirements.txt
requests
beautifulsoup4
google-generativeai
supabase
python-dotenv
PyPDF2
playwright